pyproject.toml
[project]
name = "food-image-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "streamlit>=1.48.1",
    "torchvision>=0.23.0",
    "typing>=3.10.0.0",
    "uuid>=1.30",
    "uvicorn>=0.35.0",
]

[tool.uv.workspace]
members = [
    "venv",
]
